refactor(server): modularize connector pairing boundary
